+++ R/mw-etag.R
@@ -1,9 +1,11 @@
 
-#' Middleware that add an `Etag` header to the response
+#' Middleware that add an `ETag` header to the response
+#'
+#' If the response already has an `ETag` header, then it is kept.
 #'
 #' This middleware handles the `If-None-Match` headers, and it sets the
 #' status code of the response to 304 if `If-None-Match` matches the
-#' `Etag`. It also removes the response body in this case.
+#' `ETag`. It also removes the response body in this case.
 #'
 #' @param algorithm Checksum algorithm to use. Only `"crc32"` is
 #' implemented currently.
@@ -23,8 +25,11 @@ mw_etag <- function(algorithm = "crc32") {
   }
   function(req, res) {
     do <- function(req, res) {
-      etag <- paste0("\"", crc32(res$.body), "\"")
-      res$set_header("Etag", etag)
+      etag <- res$get_header("ETag")
+      if (is.null(etag)) {
+        etag <- paste0("\"", crc32(res$.body), "\"")
+        res$set_header("ETag", etag)
+      }
       req_etag <- req$get_header("If-None-Match")
       if (!is.null(req_etag) && req_etag == etag) {
         res$.body <- NULL